Reserve 0% for root on ext4 partitions
This is to have the proper free disk space reported from XBMC,
or even df, since the default is 5%. The impact is low, since
having OpenELEC run everything as root had this limit not be
enforced at any time.
On my 3TB drive, XMBC reported 2.5TB free before the change
and 2.7TB after.