busybox-initramfs: init: cleanup, improve readability
Use consistent whitespace for all `progress "..."` lines.
Don't mount $disk if it is not set (e.g. in the installer).
Don't check for /storage/$OVERLAY_DIR, just create it.
Move the overlay functionality into an if statement.
Don't move /storage if $disk is not set (e.g. in the installer).

Signed-off-by: Alain Kalker <a.c.kalker@gmail.com>