linux-initramfs: new package, adds support for loading kernel modules
This package adds support for loading kernel modules during early boot.

The list of modules to be loaded can be specified in projects/*/options in the
variable INITRAMFS_MODULES .

The list of modules to load will be stored in initramfs /etc/modules , the
modules themselves in /lib/modules .
No subdirectories are used to ease and speed up module loading in init.
Only modules which were actually built will be added to the initramfs.

Signed-off-by: Alain Kalker <a.c.kalker@gmail.com>