scripts\checkdeps: APT-Support and checks for "sudo" on Debian ... this fixes Issue #1482
Ubuntu:
Checks if Aptitude is installed. If not, the user will be asked if he
wants to install the packages with apt-get OR if he wants to install
Aptitude and start the process again.

Debian:
Sudo is not installed by default. So the script checks if sudo is
present (maybe installed by the user befor). If not "su -c <command>"
will be used.