AppleTV: Make in-kernel config modular (but keep it available)
The in-kernel config (/proc/config.gz) is a useful way to understand how your existing kernel has been compiled, however it does use some memory to represent this information inside the kernel. However by making it optional you can still have this information if wanted, but as part of the system ramdisk (as a kernel module) rather than in memory all the time.

For AppleTV this was not (yet) enabled, for other platforms we make it modular.