AppleTV: Make sure loop devices are allocated dynamically
CONFIG_BLK_DEV_LOOP_MIN_COUNT: Number of loop devices to pre-create at init time

Static number of loop devices to be unconditionally pre-created at init time.

This default value can be overwritten on the kernel command line or with module-parameter loop.max_loop.

The historic default is 8. If a late 2011 version of losetup(8) is used, it can be set to 0, since needed loop devices can be dynamically allocated with the /dev/loop-control interface.