connman: allow user editable main.conf
In recent months I have seen a number of cases where users need to
change ‘prefer ethernet’ to ‘prefer wifi’ or where users are trying to
enable tethering for ethernet or are running other daemons that add
network interfaces that should not be auto-managed by comman. All of
these require user editing of main.conf so this change moves main.conf
to /storage where it can be tweaked under-the-hood if needed.