init: ignore ret 8 from fsck if not force checking
when /flash or /storage is on non-persistent storage, it's not
yet available at the time fsck -a -p fires, so just skip
and force fsck only if ret=4

now, if /flash or /storage is non-persistent, users get
either non-recovered fs (just as it was before) or debug
shell if the fs is fubar (bad sb and so). nobody cares.