initramfs: enable DNS if network is up
In certain situations, it is desirable for hostnames to be resolvable in
the initramfs (e.g., when a hostname is specified as the NFS server in
the "boot" or "disk" kernel parameters). If the network is up when the
init script is executed (e.g., when the "ip" kernel parameter is
specified) and a DNS server is known, write it to /etc/resolv.conf so it
can be used for DNS requests.