v4l-utils: add multi keymap for Amlogic devices
This keymap supports the common remotes (MCE, Xbox 360 etc)
plus the Odroid an Wetek Hub remote.

It is used with the meson-ir driver unless one of the other
supported keymaps was explicitly chosen in devicetree.

By default meson-ir uses the rc-empty map, Odroid C2 DT sets it
to rc-odroid and Wetek Hub DT sets it to rc-wetek-hub. As rc-maps.cfg
doesn't contain a mapping for these the amlogic multi-map will be used.

Signed-off-by: Matthias Reichl <hias@horus.com>