Revert "make: Fix glob handling (#2681)"
This reverts commit e0b6641f50f881b072b619f130914d48a05b7ef9.

Changing configure.ac requires aclocal-1.15 to be installed on the
build host as the configure script needs to be regenerated.

This breaks builds on all systems without automake-1.15, so revert
the change until a solution for the Arch make issue is found
that doesn't introduce a ton of host dependencies.

Signed-off-by: Matthias Reichl <hias@horus.com>