installer: remove unused functions, pointless case, recursive calls and faulty breaks
I suspect the recursive calls were to work around the error caused by break
being called in a case statement which itself is within a while loop - the break
causes the while loop to terminate, bringing the installer to a grinding halt... whoops.

Now, menu_main() is called only from the main process loop, which we don't exit accidentally.

I've removed an entire case statement from do_install_quick() as this function either
performs an installation, or it doesn't - the case statement is pointless.

Due to the removal of whitespace, add ?w=1 in github.com when reviewing.