ffmpeg: cleanup
disable-{ffprobe,ffplay} and enable-ffmpeg get ignored because of
disable-programs later on, so just remove them.

disable-{w32threads,dxva2} get ignored because of target-os=linux.

Hardcoded tables are a start up speed optimization in exchange for
an increased sizeon disk  of libavcodec. Its impact is negligible
due to the encode/decode cycle.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>