tools/download-cleaner: add new tool
This script selectively deletes files out of the specified download
directory to prune source tarballs that don't correspond with the
current version of its package in the tree according to the set
PROJECT, DEVICE and ARCH.

The script is also capable of building a new SOURCES dir to combine
different combinations of PROJECT, DEVICE and ARCH tarballs. This is
to move out known good tarballs to a new directory, and to delete the
old SOURCES dir once all desired tarballs have been moved.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>