v4l-utils: simplify 70-infrared.rules and fix kernel 4.20 issues
Using $id to get the rcX device name from the SUBSYSTEMS match,
as used by the upstrem v4l-utils udev rule, is easier and more
stable than using IMPORT{parent}, which needs to transmit info across
two levels (rcX->inputY->eventZ) and breaks with kernel 4.20.

Instead of running a shell script to check for the existance of
/storage/.config/rc_maps.cfg we can use the builtin udev TEST
function.

Also it's better to use ENV variables prefixed with a dot as
they are used only locally in the udev rule file and don't need
to be persisted in the udev database or exported to external tools.

Signed-off-by: Matthias Reichl <hias@horus.com>