update_retroplayer-addons: sync libretro core versions with kodi
Instead of bumping libretro cores to their current master version
use the githash from depends file in kodi game addon.

This ensures we ship the same libretro core versions as Kodi.

When a kodi game addon is bumped the libretro package is automatically
bumped, too.

If the script is invoked with the "-f" option all libretro packages
will be synced to the version specified in the kodi game addon.

Signed-off-by: Matthias Reichl <hias@horus.com>