update_retroplayer-addons: don't bump PKG_REV on game addon bumps
As we are using tags, and the addon version should match the tag
version bumping PKG_REV on kodi game addon updates isn't needed.

As with update_binary-addons bumping PKG_REV may be wanted as
we are switching from HEAD to tags and can be enabled by setting
the PKG_BUMP_REV environment variable.

PKG_REV will also be bumped the script is run with "-f" as before.

Signed-off-by: Matthias Reichl <hias@horus.com>