buildsystem: HACK: fix current status after initramfs
initramfs is currently built from within linux:target which is a problem
yet to be solved as it should be built as separately scheduled processes
just like everything else.

As we build initramfs from within linux:target, once initramfs is
finished it emits an "UNLOCK" event which is the last event for the
linux:target package until it too finishes, perhaps 5 or 10 minutes later.
During this 5-10 minute period the dashboard gives the impression that
initramfs is "stuck" in an UNLOCKed state when in fact processing continues
within linux:target, which is now the "ACTIVE" package for the process slot.

Until the initramfs scheduling problem is resolved, emit an extra "ACTIVE"
event upon returning to linux:target correcting the dashboard (and also
tools/mtstats.py elapsed time accumulation).