Lakka: change online udpater URL for nightlies
if users try to upgrade nightly build, they are offered to update to
latest stable release, which is a downgrade. a patch corrects the URL to
the nightlies server. this patch is applied only when the build is
started with LAKKA_NIGHTLY=yes, so it has no impact in other scenarios.