Lakka: LIBRETRO_CORES style change (#778)
main change is that each core in the LIBRETRO_CORES list is defined on
separate line, which makes the diff with the changes to the list
(new core/removed core) more clear.

another change is, that when removing a core from the list for specific
platform, spaces before and after the core name are added. this ensures
that no partials are removed from the list and the list does not contain
non-existent packages (e.g. to keep flycast_wince, but to remove
flycast, without leading/trailing space the list would contain
non-existent package named _wince).

and last change is in the virtual package for Rockchip - removed
beetle-psx from the PKG_DEPENDS_TARGET, as it is already in
LIBRETRO_CORES.