pulseaudio: enable speex support
this fixes the following warning:

meson.build:871: WARNING:
You do not have speex support enabled. It is strongly recommended
that you enable speex support if your platform supports it as it is
the primary method used for audio resampling and is thus a critical
part of PulseAudio on that platform.