Allwinner: Add frame inversion to correct audio channels
For some reason the mainline i2s driver inverts the frame clock
which in turn maps the audio channels incorrectly when not using
audio passthrough on multi-channel audio.

Adding frame-inversion to the device tree properties fixes the
issue but should be seen as a temporary fix whilst we investigate
the correct default setup of the i2s block.

Signed-off-by: Marcus Cooper <codekipper@gmail.com>