show_config: move more of show_config kodi specific to distro version
Split Kodi configuration messages into hardware / software sections. Hardware is how
other software is being built to accomodate Kodi needs (ie what provides video accel).
Software is the configure options to Kodi.

Kodi version just outputs the velue of $MEDIACENTER, which will always be "kodi".
Change to who the "upstream" is for the version of Kodi we're using is instead.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>