tools/ffmpeg: add gen-patches.sh script
This script creates or updates the ffmpeg patch files from
the various feature branches and record patch info (repo,
branch, gitrevs) where the changes came from in the commit
message.

The feature branch to process can be specified via command a
line argument, "all" will create/update all known branches.

By default the patches are updated to HEAD of the branches.

If only a single branch is processed the full githash to use
instead of HEAD can be specified as an optional second argument.
This can be used to drop some of the top-most commits in the
branch, eg if later changes cause issues.

The script has to be run in a local ffmpeg git tree. eg:

$ cd ~/ffmpeg-git
$ ~/libreelec-git/tools/ffmpeg/gen-patches.sh all

Additional git commit options (eg -s) can be set via the
GIT_COMMIT_ARGS environment variable

Signed-off-by: Matthias Reichl <hias@horus.com>