zstd: add configure options
legacy_level=0 disables support for packages compressed with zstd before
version 0.8 (~August 2016).

Support for the other compression libraries adds a symlink feature to the
zstd binary for it to mimic that library's regular tool. Example, a
symlink pointing to zstd named 'gzip' will cause zstd to pretend to be gzip.
The cli tools aren't used, so there is no purpose to adding this support.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>