Adding RG351P support, an Odroid Go Advance clone
This console requires a separate DTS and a USB controller config.
U-Boot cannot reliably distinguish the two consoles since the HW ID
for it is the same as OGA v1.1 (looking into the schematics did not
yeld any GPIO differences that could be used for that).

I plan to dig a bit more into u-boot, perhaps enabling USB and
looking for the USB controller could be a way to differentiate them.
In the meantime this just adds another OGA flavor, which adds little
overhead to the build system.

Another TODO is rumble support, mapped to PWM1, which is not trivial
to support in RetroArch.