tools: add script to refresh patches
This script uses quilt to refresh patches, e.g.:
./tools/refresh-patches linux

It tries to detect already applied patches and removes those.
The rest is refreshed, which means updating the hunk line offsets and
removing junk (so that the result is the same for everyone).

This makes it easy to update packages with large patch sets. It's
usable as-is without any additional git clone to rebase patches.

Note that there're alot of patches in-tree that contain multiple patches
per file. While quilt doesn't have a problem applying those, the
resulting file is a combined patch.