checkdeps: overhaul
Convert from using the indices of two arrays to establish a relationship
to using associative arrays to make the relationship direct in a
[key]=value pairing.

Move the qemu-binfmt-dep, et all, for native Rockchip builds out to the
project options file.

Don't recalculate needed deps if sudo isn't installed to add new packages;
just exit.

Rename Gentoo's python3 package to python. Specifying a slot would be how to
state a specific version of python on Gentoo. Regardless, python is a hard dep
of Gentoo's package manager so this will never come up in practice.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>