Move kernel overlays setup to init
When running kernel-overlays-setup as a systemd service we have
a short window where /lib/firmware isn't available.

This causes drm edid override to occasionally fail if the edid file
is requested between switching to sysroot and completion of the
kernel-overlays-setup script.

Setting up firmware and module overlays in init, before switching
to sysroot, closes the gap and also has the benefit that we no longer
need to care about the rather tedious systemd unit ordering.

Signed-off-by: Matthias Reichl <hias@horus.com>