buildsystem: use an update lock when updating image and sysroot (#1168)
* buildsystem: add is_sequential_build() helper

* buildsystem: use an update lock when updating image and sysroot

cp (and potentially mkdir -p) are not atomic, and we have seen situations where two packages
concurrently copying the same file (eg. the udev rule for xf86-video-nvidia and
xf86-video-nvidia-legacy) will succeed for one package but the other package fails with
a "file exists" error (as the file didn't exist when it checked, but does exist when it
actually copies the file). Not even cp -f will avoid this issue.

There are several workarounds, but the most practical (and general) solution is to ensure
sequential updates of the image and shared sysroot directories.

Co-authored-by: MilhouseVH <milhouseVH.github@nmacleod.com>