linux: set DTC_FLAGS as environment variable instead of make argument
Setting DTC_FLAGS as a make argument overrides the additional dtc flags
which are added in kernel's scripts/Makefile.lib.

This results in lots of noisy warnings when building RPi DT and overlays
and could be potentially problematic if kbuild adds important dtc flags.

Pass in DTC_FLAGS as an environment variable so that kbuild can
adjust DTC_FLAGS as needed.

Signed-off-by: Matthias Reichl <hias@horus.com>