Adding initial support for Gameforce CHI console (#1182)
This is essentially an OGA clone with a bigger (640x480) non-rotated
screen. Requires a joypad kernel patch (a copy of OGA's + small bits) and a new
DTB.

Changed a bit the way splash screens work, now you can specify proper
resolutions (instead of just vertical res) so that it works well for
cases like rotated screens.

Tweaked Retroarch to ship both drivers (DRM + GO2 DRM) and the GO2
driver to inhibit itself in case the console has a non-rotated display.
This should work for any console as long as the resolution is reported
correctly. The regular DRM driver should have 1 frame less latency!

What's not working: Console is still rotated (minor issue) and kernel
panics on poweroff :) Debugging it, but I'll need to solder the UART for
that.