Rockchip: add mapping for mainline device trees
This adds a mapping for device trees which differ between vendor and mainline kernel and will therefore allow updates to new releases.
The old ones will be replaced by the new, or at least most appropriate (in case of RK3328 box devices), devicetrees in mainline kernel.

Besides a warning will be shown now if the device tree that the installation uses does not exist in the update tarball.