rust: update to 1.54.0 and enable host cross compiles
Until target-applies-to-host is incorporated into stable this
"target-applies-to-host" option is required for a matching
host-target triple to be compiled by the cross compiler.

Read more here:
https://doc.rust-lang.org/cargo/reference/unstable.html#target-applies-to-host

When rust-lang/cargo merges the following and it is released - then this "rust"
package will need to be revisted if it is bumped.
- Stabilize target-applies-to-host feature. #9753

The following are the rust-lang/cargo PRs that made this possible:
- Configure hosts separately from targets when --target is specified. #9322
- Tracking Issue for target-applies-to-host #9453