RPi4: temporarily limit framebuffer size to HD
RPi firmware allocates a 4k framebuffer in gpu mem when 4kp60 is
enabled which leads to H264 decoding failing because it's running
out of GPU memory.

This framebuffer isn't used/needed in LE at all, so limit it's size
to HD (which the firmware does in non-4kp60 setups) until the KMS driver
and the firmware support freeing the orphaned framebuffer.

Signed-off-by: Matthias Reichl <hias@horus.com>