duckstation: use specific version of the binary core (#1464)
* duckstation: use specific version of the core

now it is possible to use specific version of the core based on the
published changelog and regular fetching of the binaries to the Lakka
mirror server

* libretro_update.sh: do not update PKG_VERSION to blank value

when the ls-remote command returns an error, do not update PKG_VERSION
with a blank value. happens for packages that do not use a git
repository (e.g. duckstation at the moment)