u-boot: use setup_pkg_config_host for build
Issue was that pkg-config picks wrong include folders. That is because
in u-boot case, pkg-config is configured to pick up target (ARM)
includes and libraries, because it's called from make_target(). Simple
solution to avoid patching sources and adding CONFIG_TOOLS_LIBCRYPTO
line to each and every config file is to add setup_pkg_config_host at
the beginning of make_target(). Include the dependency to
openssl:host.