mkimage: rework FAT32 population
In case SYSTEM_SIZE was higher than 1024, mcopy would fail without the
failure being caught. The failure was probably due to invalid geometry
(number of sectors per track too high?). After discussion on IRC with
vpeter we came to this solution.

* extract the FAT32 partition from the image
* format the extracted partition without any geometry parameters
  (mformat is able to determine the geometry by itself)
* install/copy files to the extracted partition
* verify the partition
* merge the partition back to the image

The extraction of the FAT32 partition is also needed for the OVA image
in case of Generic project.

Errors while mformat/mcopy/mmd will now be caught and image creation
will be terminated with an error message.

Removed alias for mformat, as it is used only once. Added `-o` to alias
for mcopy, just in case, as the output is now redirected, in case a file
would already exist, the prompt would be not shown and image creation
would just hang (ctrl-c does not work, had to use kill -9 on mcopy
during my testing when I forgot the `-o` switch).

For OVA full path to toolchain is not needed, as the alias has already
the partition in it, not the image.